Added some error and exception handling, added more unit tests

+++ gethoney/main.py
@@ -1,4 +1,4 @@
-from fastapi import Depends, FastAPI
+from fastapi import Depends, FastAPI, HTTPException
 
 from gethoney.crud import Database
 from gethoney.models import Honeypot
@@ -14,16 +14,16 @@ def get_db():
 
 @app.get("/honeypots/{id}", status_code=200)
 def get_honeypot(id: int, db: Database = Depends(get_db)):
-    print(id)
-    return db.read_honeypot(id)
+    honeypot = db.read_honeypot(id)
+    if not honeypot:
+        raise HTTPException(status_code=404, detail="Honeypot not found.")
+    else:
+        return honeypot
 
 
-@app.get("/honeypots/")
+@app.get("/honeypots/", status_code=200)
 def list_honeypots(db: Database = Depends(get_db)):
     return db.list_honeypots()
-
-
-#  TODO: Change from list of Honeypots to single Honeypot
 
 
 @app.post("/honeypots/", status_code=201)
@@ -34,9 +34,18 @@ def create_honeypot(honeypot: Honeypot, db: Database = Depends(get_db)):
 
 @app.delete("/honeypots/{honeypot_id}", status_code=200)
 def delete_honeypot(honeypot_id: int, db: Database = Depends(get_db)):
-    return db.delete_honeypot(honeypot_id)
+    honeypot = db.read_honeypot(honeypot_id)
+    if not honeypot:
+        raise HTTPException(status_code=404, detail="Honeypot not found.")
+    else:
+        db.delete_honeypot(honeypot_id)
+        return honeypot
 
 
 @app.put("/honeypots/{honeypot_id}", status_code=200)
 def update_honeypot(honeypot_id: int, honeypot: Honeypot, db: Database = Depends(get_db)):
-    return db.update_honeypot(honeypot, honeypot_id)
+    update = db.update_honeypot(honeypot, honeypot_id)
+    if not update:
+        raise HTTPException(status_code=404, detail="Honeypot not found.")
+    else:
+        return update
